requirements.txt
aiogram==3.10.0
python-dotenv==1.0.0
aiohttp>=3.9.0
google-api-python-client==2.95.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.2.0
Pillow==10.0.0
aiofiles~=23.2.1
